moving currency to diff tab
