+++ crypto_balance/lib/entities/currency.dart
